[presign] Refactor fees and trade summary (#1313)
# Summary

This is the first of several PRs for adding Smart Contract support. 

This first PR scope is just three things:
* Refactoring the TradeSummary into reusable components
* Adds to the component that display the fee some parametrisation
* Use this parametrization to:
   * Remove the label where it says the gas costs are included
   * Use another tooltip text meant for wallets that need to pre-sign

![image](https://user-images.githubusercontent.com/2352112/130789518-fc647344-7bed-41e7-ad56-7565e45c3da5.png)


## Out of the scope
* **Final design on how to show the tx Fee** The goal of this PR is not to make the user fully aware when he has to pay for the gas costs, but instead just refactor and parametrize in preparation for next PRs (so we can make review easier). This part still require some research on how is best to make the user aware
* **Detect when we should use pre-sign instead of offchain signing**
* **SC pre-signing**: Will come in future PRs


## Changes
`TradeSummary` was an old Mod that fell out of sync after doing so many changes. I think this happened in V3.

I think that, it has changed so much that we can start to own this one, and make it a normal component. 

For that, this PR:
* Deletes `TradeSummaryMod`, now main logic of the component is in `TradeSummary`
* To make `TradeSummary` smaller and easier to mantain, the reusable rows where moved to it's own file: 
   * `RowFee`: Display the fee
   * `RowReceivedAfterSlippage`: Display the amounts received after fee
   * `RowSlippage`: Show the slippage
 * Reuse the above components, so they are also used in the confirmation modal you are shown before trading. This reusing/refactor started in #1149, but this PR continued the refactor and introduced `RowFee`.

## Bonus
Since now `RowFee` has been refactored and reused, a nice bonus of this PR is that we get USD estimation in the confirmation modal, which we don't have in production right now 🎉

Also, of course reduce some duplication we had!

![image](https://user-images.githubusercontent.com/2352112/130789931-0e4255b1-2ba7-4b37-8697-66c88eb494a4.png)

![image](https://user-images.githubusercontent.com/2352112/130789992-9db2c8e6-8f71-4e07-b4a8-73b0a9d1fbf0.png)

# Help for reviewers
I tried to keep the code and logic as similar as possible as before. So even if you see many changes, the original code is preserved.

The only parts that were changed a bit, is the part where I choose the tooltip and fee label sufix depeding on `allowsOffchainSigning`


# To Test

1. Try to review that the fee is displayed correctly:
- When we preview the trade in the trade widget
- When we preview the trade in the confirmation modal
- When we preview the trade in the trade widget, and the fee is higher than the sold amount

Ignore for now that now, the label "(incl. gas costs)" is now gone. Future PRs will define exactly what we show in each case.
